[pythonic config] Restructure Pythonic config & resources into modules
