+++ python_modules/libraries/dagstermill/dagstermill/asset_factory.py
@@ -17,7 +17,7 @@ from dagster import (
     asset,
 )
 from dagster._config.pythonic_config import Config, infer_schema_from_config_class
-from dagster._config.pythonic_config.utils import safe_is_subclass
+from dagster._config.pythonic_config.inheritance_utils import safe_is_subclass
 from dagster._core.definitions.events import CoercibleToAssetKey, CoercibleToAssetKeyPrefix
 from dagster._core.definitions.utils import validate_tags
 from dagster._core.execution.context.compute import OpExecutionContext
